+++ API-Editor/ComponentInformation.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using JetBrains.Annotations;
 using UnityEngine;
+using Object = UnityEngine.Object;
 
 namespace Anatawa12.AvatarOptimizer.API
 {
@@ -47,6 +48,9 @@ namespace Anatawa12.AvatarOptimizer.API
             ComponentMutationsCollector collector) =>
             CollectMutations((TComponent)component, collector);
 
+        internal override void ApplySpecialMappingInternal(Component component, MappingSource collector) =>
+            ApplySpecialMapping((TComponent)component, collector);
+
         /// <summary>
         /// Collects runtime mutations by <see cref="component"/>.
         /// You have to call <see cref="collector"/>.<see cref="ComponentMutationsCollector.ModifyProperties"/>
@@ -68,7 +72,18 @@ namespace Anatawa12.AvatarOptimizer.API
         protected virtual void CollectMutations(TComponent component, ComponentMutationsCollector collector)
         {
         }
-        
+
+        /// <summary>
+        /// Applies some property mapping to your component.
+        /// For object replacements, AAO processes automatically so you don't have to implement that in this method.
+        /// </summary>
+        /// <param name="component">The component.</param>
+        /// <param name="mappingSource">The mapping source</param>
+        [PublicAPI]
+        protected virtual void ApplySpecialMapping(TComponent component, MappingSource mappingSource)
+        {
+        }
+
         // Important note for future AAO developer
         // 1. You MUST NOT add abstract method onto this class
         // 2. When you added some virtual methods onto this class, implementer MAY NOT implement that method
@@ -104,15 +119,29 @@ namespace Anatawa12.AvatarOptimizer.API
         public abstract void MarkEntrypoint();
 
         /// <summary>
-        /// Marks this component as Behaviour component, which means has some effects to components in the avatars.
+        /// Marks this component as HeavyBehaviour component, which means the component uses some resources while
+        /// enabled but doesn't eat if not enabled and AAO can (almost) freely change enablement of the component.
         /// When you mark some components Behaviour component, Avatar Optimizer will generate animation that disables
         /// the component when entrypoint is not active / enabled.
         ///
-        /// If your component is not marked as Behaviour, enable-ness / activeness will not changed by Avatar Optimizer.
         /// If your component have some runtime load and can be skipped if your component is not needed by all
-        /// EntryPoint components, you should mark your component as Behaviour for runtime-load optimization.
+        /// enabled EntryPoint components, you should mark your component as Behaviour for runtime-load optimization.
         ///
-        /// For example, VRCPhysBone and Constraints are marked as Behaviour.
+        /// For example, VRCPhysBone and Constraints are marked as HeavyBehaviour.
+        /// </summary>
+        [PublicAPI]
+        public abstract void MarkHeavyBehaviour();
+
+        /// <summary>
+        /// Marks this component as Behaviour component, which means the activeness of the component has meaning.
+        /// When you mark your some components Behaviour component, Avatar Optimizer will never change activeness of
+        /// the component.
+        ///
+        /// If your component is not a Behaviour component, AAO may change enablement of the component.
+        ///
+        /// NOTE: In AAO 1.6.0, AAO will not change enablement of non-Behaviour components but in the feature releases,
+        /// AAO may change enablement of non-Behaviour components to change enablement of HeavyBehaviour components
+        /// effectively.
         /// </summary>
         [PublicAPI]
         public abstract void MarkBehaviour();
@@ -180,5 +209,66 @@ namespace Anatawa12.AvatarOptimizer.API
 
         [PublicAPI]
         public abstract void ModifyProperties([NotNull] Component component, [NotNull] IEnumerable<string> properties);
+
+        [PublicAPI]
+        public void ModifyProperties([NotNull] Component component, [NotNull] string[] properties) =>
+            ModifyProperties(component, (IEnumerable<string>) properties);
+    }
+
+    public abstract class MappingSource
+    {
+        internal MappingSource()
+        {
+        }
+
+        [PublicAPI]
+        public abstract MappedComponentInfo<T> GetMappedComponent<T>(T component) where T : Component;
+
+        [PublicAPI]
+        public abstract MappedComponentInfo<GameObject> GetMappedGameObject(GameObject component);
+    }
+
+    public abstract class MappedComponentInfo<T> where T : Object
+    {
+        internal MappedComponentInfo()
+        {
+        }
+
+        /// <summary>
+        /// The mapped component (or GameObject).
+        /// The component may be removed without mapped component.
+        /// If there are not mapped component, this will be null.
+        ///
+        /// Even if the component is removed without mapped component,
+        /// each animation property can be mapped to another component.
+        /// </summary>
+        [PublicAPI]
+        public abstract T MappedComponent { get; }
+
+        /// <summary>
+        /// Maps animation property name to component and MappedPropertyInfo.
+        /// If the property is not removed, returns true and <paramref name="found"/> is set.
+        /// If the property is removed, returns false and <paramref name="found"/> will be default.
+        /// </summary>
+        /// <param name="property">The name of property will be mapped</param>
+        /// <param name="found">The result parameter</param>
+        /// <returns>Whether if the property is successfully mapped or removed</returns>
+        [PublicAPI]
+        public abstract bool TryMapProperty(string property, out MappedPropertyInfo found);
+    }
+
+    public readonly struct MappedPropertyInfo
+    {
+        [PublicAPI]
+        public Object Component { get; }
+
+        [PublicAPI]
+        public string Property { get; }
+
+        internal MappedPropertyInfo(Object component, string property)
+        {
+            Component = component;
+            Property = property;
+        }
     }
 }