+++ Editor/ContextExtensions.cs
@@ -1,4 +1,5 @@
 using System;
+using Anatawa12.AvatarOptimizer.AnimatorParsers;
 using Anatawa12.AvatarOptimizer.Processors;
 using Anatawa12.AvatarOptimizer.Processors.SkinnedMeshes;
 using JetBrains.Annotations;
@@ -47,5 +48,43 @@ namespace Anatawa12.AvatarOptimizer
 
         public static void RecordRemoveProperty([NotNull] this BuildContext context, Component from, string oldProp) =>
             GetMappingBuilder(context).RecordRemoveProperty(from, oldProp);
+
+        public static AnimationComponentInfo GetAnimationComponent([NotNull] this BuildContext context,
+            ComponentOrGameObject component) =>
+            GetMappingBuilder(context).GetAnimationComponent(component);
+        
+        public static bool? GetConstantValue([NotNull] this BuildContext context, ComponentOrGameObject obj, string property, bool currentValue)
+        {
+            if (!context.GetAnimationComponent(obj).TryGetFloat(property, out var prop))
+                return currentValue;
+
+            switch (prop.State)
+            {
+                case AnimationFloatProperty.PropertyState.ConstantAlways:
+                    return FloatToBool(prop.ConstValue);
+                case AnimationFloatProperty.PropertyState.ConstantPartially:
+                    var constValue = FloatToBool(prop.ConstValue);
+                    if (constValue == currentValue) return currentValue;
+                    return null;
+                case AnimationFloatProperty.PropertyState.Variable:
+                    return null;
+                case AnimationFloatProperty.PropertyState.Invalid:
+                default:
+                    throw new ArgumentOutOfRangeException();
+            }
+
+            bool? FloatToBool(float f)
+            {
+                switch (f)
+                {
+                    case 0:
+                        return false;
+                    case 1:
+                        return true;
+                    default:
+                        return null;
+                }
+            }
+        }
     }
 }