+++ Editor/Utils/Utils.VRCSDK.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 using UnityEngine;
 using VRC.Dynamics;
@@ -25,5 +26,8 @@ namespace Anatawa12.AvatarOptimizer
                         queue.Enqueue(child);
             }
         }
+
+        // https://creators.vrchat.com/avatars/#proxy-animations
+        public static bool IsProxy(this AnimationClip clip) => clip.name.StartsWith("proxy_", StringComparison.Ordinal);
     }
 }